requirements.txt
alpaca-py>=0.28.0
pandas>=2.1.0
numpy>=1.26.0
python-dotenv>=1.0.0
PyYAML>=6.0
